requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
discord-webhook==0.17.0
markdownify==0.11.2
python-dotenv==0.20.0
requests==2.25.1